rename buckets.py to routes.py
