+++ app.py
@@ -1,3 +1,3 @@
 from api.imports import app
-from api.v1.buckets import *
+from api.v1.routes import *
 app.run(host='127.0.0.1', port=1111)
